Added message to debug serial port problems
The error message is printed to the console only if the serial port is not connected. If the serial port RX pin is not properly connected, this message will be printed (and also if the serial port is not present at all). This should shorten debug time in case of future errors.